+++ frontend/src/scenes/groups/cards/GroupCard.tsx
@@ -1,0 +1,3 @@
+export function GroupCard({ children }: { children: React.ReactNode }): JSX.Element {
+    return <div className="border rounded bg-surface-primary py-2 px-1">{children}</div>
+}
